Fix 32-bit time_t narrowing in undo_time (CID 466054)
undo_time() held target/closest/closest_start/val -- all of which can
carry a time_t-range value in "seconds since epoch" mode -- in plain
int locals, narrowed via (int) casts from b_u_time_cur, os_time(), and
uh_time (the latter read unvalidated from a .un~ undo file).

This is not just a post-2038 problem: INT_MAX minus the current epoch
time is presently on the order of 12 years, so ":later {N}d" already
overflows int today for a large-but-plausible N (":later 4527d" and
up, as of 2026) -- no crafted file or clock manipulation needed. On
overflow, the wrapped target corrupts the "which undo header is
closest" ranking, so the wrong point in history can get selected.

Widened target/closest/closest_start/val to int64_t and corrected
their assignments accordingly. One (int) cast is kept, where target
flows into a sequence number after the code switches from time-based
to seq-based search (guarded by "above", which review confirmed is
only ever set immediately after target is reassigned to a sequence
number) -- and that cast is safe independent of that invariant too,
since the comparison it depends on (uhp->uh_seq == target) now
promotes the int uh_seq up to int64_t, so no truncating match is
possible.

Also confirmed, out of scope for this CID and left as-is: three
sibling sites (b_u_seq_cur + step, b_u_save_nr_cur + step, twice) have
the same class of int-arithmetic overflow on an adversarial count, but
are not time-carrying values and were not observed to produce wrong
behavior in testing -- worth a follow-up look, not folded in here.

Verified via a 2,000,000-case differential arithmetic harness (0
divergences from the pre-fix behavior on legitimate, non-overflowing
input) and direct exercise of g-/g+, :earlier/:later in both time and
file-count modes, plus the file-write-count overflow case, against the
built binary. The existing oldtest suite's own time-based
":earlier {N}s" cases are skipped upstream (missing test_settime()
support), so they provide no coverage of this code path either way.

+++ src/nvim/undo.c
@@ -1991,8 +1991,12 @@ void undo_time(int step, bool sec, bool file, bool absolute)
     u_oldcount = -1;
   }
 
-  int target;
-  int closest;
+  // "target" and "closest" hold a timestamp when "dosec" is set, otherwise a
+  // file-write counter or an undo sequence number.  Timestamps are time_t
+  // values, which are wider than int, so use int64_t to hold either kind
+  // without truncating.
+  int64_t target;
+  int64_t closest;
   u_header_T *uhp = NULL;
   bool dosec = sec;
   bool dofile = file;
@@ -2006,7 +2010,7 @@ void undo_time(int step, bool sec, bool file, bool absolute)
     closest = -1;
   } else {
     if (dosec) {
-      target = (int)curbuf->b_u_time_cur + step;
+      target = (int64_t)curbuf->b_u_time_cur + step;
     } else if (dofile) {
       if (step < 0) {
         // Going back to a previous write. If there were changes after
@@ -2049,7 +2053,7 @@ void undo_time(int step, bool sec, bool file, bool absolute)
       closest = -1;
     } else {
       if (dosec) {
-        closest = (int)(os_time() + 1);
+        closest = (int64_t)os_time() + 1;
       } else if (dofile) {
         closest = curbuf->b_u_save_nr_last + 2;
       } else {
@@ -2060,7 +2064,7 @@ void undo_time(int step, bool sec, bool file, bool absolute)
       }
     }
   }
-  int closest_start = closest;
+  int64_t closest_start = closest;
   int closest_seq = curbuf->b_u_seq_cur;
   int mark;
   int nomark = 0;  // shut up compiler
@@ -2093,9 +2097,9 @@ void undo_time(int step, bool sec, bool file, bool absolute)
 
     while (uhp != NULL) {
       uhp->uh_walk = mark;
-      int val = dosec ? (int)(uhp->uh_time)
-                      : dofile ? uhp->uh_save_nr
-                               : uhp->uh_seq;
+      int64_t val = dosec ? (int64_t)uhp->uh_time
+                          : dofile ? uhp->uh_save_nr
+                                   : uhp->uh_seq;
 
       if (round == 1 && !(dofile && val == 0)) {
         // Remember the header that is closest to the target.
@@ -2267,7 +2271,9 @@ target_zero:
         // Stop when going backwards in time and didn't find the exact
         // header we were looking for.
         if (uhp->uh_seq == target && above) {
-          curbuf->b_u_seq_cur = target - 1;
+          // "above" is only set once "target" holds a sequence number, so it
+          // fits in an int here.
+          curbuf->b_u_seq_cur = (int)(target - 1);
           break;
         }
